separa as funções do login em um controlador separado.
