+++ app/src/main/AndroidManifest.xml
@@ -12,8 +12,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MoviesActivity"></activity>
-        <activity android:name=".LoginActivity">
+        <activity android:name=".activities.MoviesActivity"></activity>
+        <activity android:name=".activities.login.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
