+++ app/src/main/java/com/androidchallengebtg/helpers/connection/Connection.java
@@ -4,9 +4,9 @@ import android.content.Context;
 import android.util.Log;
 
 import com.android.volley.Request;
+import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
-import com.android.volley.Response;
 import com.androidchallengebtg.R;
 import com.androidchallengebtg.application.ApplicationBTG;
 import com.androidchallengebtg.helpers.storage.PrefManager;
