+++ app/src/main/java/com/androidchallengebtg/helpers/connection/ConnectionListener.java
@@ -1,7 +1,5 @@
 package com.androidchallengebtg.helpers.connection;
 
-import com.android.volley.VolleyError;
-
 import org.json.JSONObject;
 
 public interface ConnectionListener {
